BOR Disabled
WDT and BOR are vital reset components to the MCU. WDT will randomly resets the program to avoid system crashes and failures, BOR will act as a set point that repeatedly checks out the MCU VDD to maintain the steady voltage. But when it comes to the most scenarios, These will cause unnecessary crashes on the MCU program execution. Therefore WDT,BOR,MCLR and STACK_OVER/UNDER_FLOW_RESETS are disabled.